Refactor ProfileStore/ProfileDatabase
